javadoc for ConnectionServer; shutdown synchronization; global close
